move mock container function out
